Move Diagnostics and Dialog warnings to Advanced settings pane
The divider and everything below it in General (Diagnostics toggle +
log retention picker + Manage Warnings button) now lives in Advanced,
after the Software update row, separated by the same divider. General
is now startup / document save / conflict resolution only.
